- Update on Revolut Api - implementation of unit test and business case. - Code refactoring

+++ src/RevolutAPI/RevolutAPI/Models/Account/GetAccountResp.cs
@@ -1,8 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-using Newtonsoft.Json;
 
 namespace RevolutAPI.Models.Account
 {
@@ -10,7 +6,7 @@ namespace RevolutAPI.Models.Account
     {
         public string Id { get; set; }
         public string Name { get; set; }
-        public double Balance { get; set; }
+        public decimal Balance { get; set; }
         public string Currency { get; set; }
         public string State { get; set; }
         public bool Public { get; set; }
